+++ package.json
@@ -9,7 +9,6 @@
   "license": "MIT",
   "devDependencies": {
     "@11ty/eleventy": "^0.11.0",
-    "@11ty/eleventy-cache-assets": "^2.0.2",
     "axios": "^0.19.2",
     "chalk": "^4.1.0",
     "dotenv": "^8.2.0",
