+++ src/_data/blogposts.js
@@ -1,80 +1,71 @@
 // required packages
-const fetch = require("node-fetch");
+const axios = require("axios");
+const path = require("path");
+const chalk = require("chalk");
+const flatCache = require("flat-cache");
 
-// max number of records to fetch per query
-const recordsPerQuery = 1;
+// Config
+const ITEMS_PER_REQUEST = 1;
+const CACHE_KEY = "blogposts";
+const CACHE_FOLDER = path.resolve("./_datacache");
+const CACHE_FILE = "blogposts.json";
 
-// DatoCMS token
-const token = process.env.DATOCMS_TOKEN;
+/**
+ * Request blogposts from DatoCMS
+ * @param {Int} skipRecords - number or records to skip
+ * @return {Array} - API data using Axios
+ */
+async function requestBlogposts(skipRecords = 0) {
+  const request = await axios({
+    url: "https://graphql.datocms.com/",
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+      Accept: "application/json",
+      Authorization: `Bearer ${process.env.DATOCMS_TOKEN}`,
+    },
+    data: {
+      query: `{
+        _allBlogpostsMeta {
+          count
+        }
+        allBlogposts (
+          skip: ${skipRecords}
+          first: ${ITEMS_PER_REQUEST}
+          filter: {
+            _status: {eq: published}
+          }
+        )
+        {
+          id
+          _createdAt
+          title
+          slug
+          intro
+          body(markdown: true)
+          image {
+            url
+            alt
+          }
+          relatedBlogs {
+            id
+          }
+        }
+      }`,
+    },
+  });
 
-// get blogposts
-// see https://www.datocms.com/docs/content-delivery-api/first-request#vanilla-js-example
-async function getAllBlogposts() {
-  // number of records to skip (we start at 0)
-  let recordsToSkip = 0;
+  const data = await request.data;
+  return data;
+}
 
-  // keep querying
-  let makeNewQuery = true;
-
-  // Blogposts array
-  let blogposts = [];
-
-  while (makeNewQuery) {
-    // initiate fetch
-    const dato = await fetch("https://graphql.datocms.com/", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-        Authorization: `Bearer ${token}`,
-      },
-      body: JSON.stringify({
-        query: `{
-            allBlogposts(
-              first: ${recordsPerQuery},
-              skip: ${recordsToSkip},
-              orderBy: _createdAt_DESC,
-              filter: {
-                _status: {eq: published}
-              }
-            )
-            {
-              id
-              title
-              slug
-              intro
-              body(markdown: true)
-              _createdAt
-              image {
-                url
-                alt
-              }
-              relatedBlogs {
-                id
-              }
-            }
-          }`,
-      }),
-    });
-
-    // store the JSON response when promise resolves
-    const response = await dato.json();
-
-    // update our blogpost array with the data from the JSON response
-    blogposts = blogposts.concat(response.data.allBlogposts);
-
-    // prepare for next query
-    recordsToSkip += recordsPerQuery;
-
-    // check if we got back less than the records we fetch per query
-    // if yes, stop querying
-    if (response.data.allBlogposts.length < recordsPerQuery) {
-      makeNewQuery = false;
-    }
-  }
-
-  // format blogposts objects
-  const blogpostsFormatted = blogposts.map((item) => {
+/**
+ * Format Data coming from API
+ * @param {Array} data - array of objects to format
+ * @return {Array} - array of formatted and sorted objects
+ */
+function formatData(data) {
+  const dataFormatted = data.map((item) => {
     return {
       id: item.id,
       date: item._createdAt,
@@ -88,9 +79,72 @@ async function getAllBlogposts() {
     };
   });
 
-  // return formatted blogposts
-  return blogpostsFormatted;
+  dataFormatted.sort((a, b) => a.date - b.date);
+
+  return dataFormatted;
 }
 
-// export for 11ty
+/**
+ * Get all blogposts
+ * - check if we have a cache
+ * - if not ping api and create cache
+ */
+async function getAllBlogposts() {
+  // load cache
+  const cache = flatCache.load(CACHE_FILE, CACHE_FOLDER);
+  const cachedItems = cache.getKey(CACHE_KEY);
+
+  // if we have a cache, return cached data
+  if (cachedItems) {
+    console.log(chalk.yellow("Blogposts from cache"));
+    return cachedItems;
+  }
+
+  // if we do not, make queries
+  console.log(chalk.yellow("Blogposts from API"));
+
+  // variables
+  let requests = [];
+  let apiData = [];
+  let additionalRequests = 0;
+
+  try {
+    // make first request and push results
+    const request = await requestBlogposts();
+    apiData.push(...request.data.allBlogposts);
+    // check if we need additonal requests to get everything
+    const totalItems = request.data._allBlogpostsMeta.count;
+    additionalRequests = Math.ceil(totalItems / ITEMS_PER_REQUEST) - 1;
+  } catch (err) {
+    console.error(err);
+  }
+
+  // create additional requests
+  for (let i = 1; i <= additionalRequests; i++) {
+    let start = i * ITEMS_PER_REQUEST;
+    const request = requestBlogposts(start);
+    requests.push(request);
+  }
+
+  // additional requests in parallel
+  try {
+    const allResults = await Promise.all(requests);
+    allResults.map((result) => {
+      apiData.push(...result.data.allBlogposts);
+    });
+  } catch (err) {
+    console.error(err);
+  }
+
+  // format data
+  const formattedData = formatData(apiData);
+
+  // set and save cache
+  cache.setKey(CACHE_KEY, formattedData);
+  cache.save();
+
+  // return API items
+  return formattedData;
+}
+
 module.exports = getAllBlogposts;
