+++ src/_includes/layouts/base.njk
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>{{ htmlTitle }}{% if htmlTitle %} - {% endif %}{{ site.title }}</title>
+  <link rel="stylesheet" href="/assets/css/main.css">
 </head>
 <body>
   {% block content %}{% endblock %}
