+++ src/assets/css/main.css
@@ -1,0 +1,4 @@
+.u-fluidimg {
+  max-width: 100%;
+  vertical-align: top;
+}
